refactor: split large components into subfolders
inquiry-form/ — extract ServiceDropdown and SuccessCard into their
own files; orchestrator slimmed to form logic only

background/ — extract canvas drawing functions into draw.ts;
background.tsx becomes a lean setup/teardown component
