+++ app/kontakt/page.tsx
@@ -1,5 +1,5 @@
 import SectionLabel from '@/components/cosmos/section-label';
-import InquiryForm from '@/components/sections/inquiry-form';
+import InquiryForm from '@/components/sections/inquiry-form/inquiry-form';
 import ContactChannels from '@/components/sections/contact-channels';
 import { services } from '@/lib/services/data';
 import styles from './page.module.css';
