+++ app/layout.tsx
@@ -3,7 +3,7 @@ import { JetBrains_Mono } from 'next/font/google';
 import './globals.css';
 import Header from '@/components/layout/header';
 import Footer from '@/components/layout/footer';
-import CosmosBackground from '@/components/cosmos/background';
+import CosmosBackground from '@/components/cosmos/background/background';
 import CosmosCursor from '@/components/cosmos/cursor';
 import { siteUrl } from '@/lib/config';
 import { Analytics } from '@vercel/analytics/next';
